Change add approval process wizard to modal
