--- src/test/smart-components/workflow/add-groups/summary-content.test.js
@@ -1,51 +1,0 @@
-import React from 'react';
-import { mount } from 'enzyme';
-import toJson from 'enzyme-to-json';
-import SummaryContent from '../../../../smart-components/workflow/add-groups/summary-content';
-import rendererContext from '@data-driven-forms/react-form-renderer/dist/cjs/renderer-context';
-import { IntlProvider } from 'react-intl';
-
-describe('<SummaryContent />', () => {
-  let formData;
-
-  it('should render correctly without groups', () => {
-    formData = {
-      name: 'Group summary name',
-      description: 'Group summary description',
-      wfGroups: []
-    };
-
-    const wrapper = mount(
-      <IntlProvider locale="en">
-        <rendererContext.Provider value={ { formOptions: {
-          getState: () => ({ values: formData })
-        }} }>
-          <SummaryContent />
-        </rendererContext.Provider>
-      </IntlProvider>
-    );
-    expect(toJson(wrapper)).toMatchSnapshot();
-  });
-
-  it('should render correctly with groups', () => {
-    formData = {
-      name: 'Group summary name',
-      description: 'Group summary description',
-      wfGroups: [
-        { value: '1', label: 'Group 1' },
-        { value: 'G2', label: 'Group 2' }
-      ]
-    };
-
-    const wrapper = mount(
-      <IntlProvider locale="en">
-        <rendererContext.Provider value={ { formOptions: {
-          getState: () => ({ values: formData })
-        }} }>
-          <SummaryContent />
-        </rendererContext.Provider>
-      </IntlProvider>
-    );
-    expect(toJson(wrapper)).toMatchSnapshot();
-  });
-});
